1.19.4 (incredible)

+++ src/main/java/org/bleachhack/module/mods/Nofall.java
@@ -34,7 +34,7 @@ public class Nofall extends Module {
 
 		if (mc.player.fallDistance > 2.5f && getSetting(0).asMode().getMode() == 1 &&
 				mc.world.getBlockState(mc.player.getBlockPos().add(
-						0, -1.5 + (mc.player.getVelocity().y * 0.1), 0)).getBlock() != Blocks.AIR) {
+						0, (int) (-1.5 + (mc.player.getVelocity().y * 0.1)), 0)).getBlock() != Blocks.AIR) {
 			mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.OnGroundOnly(false));
 			mc.player.networkHandler.sendPacket(new PlayerMoveC2SPacket.PositionAndOnGround(
 					mc.player.getX(), mc.player.getY() - 420.69, mc.player.getZ(), true));
